Move the SSH executor to ssh_executor.py
